Fix --to="all" for outputs defined in YAML header and chunk labels
When multiple outputs were defined in the yaml header, this
was ignored completly (bug in regex for YAML header).

Chunk labels are now supported for naming the plot files coming
from a particular chunk. The needed a new strategy for the
ExecutionContext, which has now a lifespan for each
document-output_format combination to be useable as "unnamed chunk
counter".