Add configuration class for final output
Add a FinalOutputConfiguration class which describes the final output
and let the rest of the process use it.

This also enables more features of yaml description of output,
although most of the options which knitr implementes are not yet
done.

Another aspect of this changes enables special handling for the
supported / prefered image filetypes per output (e.g. png for html,
pdf for pfd).