Fix not honoring `include=False` code when `eval=False`
`eval=False` adds code to output regardless of the value of `include`.
It seems like `include=False` implies `echo=False` and
`results="hide"`, so do this.

Also amend help text for `include` argument, to remove statement
that chunk will not be run. This should depend only on the `eval`
argument, according to the example document.